feat: 토큰 만료 시 재로그인 안내 모달 표시 (App.jsx로 이동)
- 세션 만료 모달을 App.jsx로 이동해 Layout과 분리
- Layout.jsx는 feat/mypage 네비 변경사항(내 펫·마이페이지 우측 정렬) 유지
- axios.js: refresh 실패 시 session-expired 이벤트 발생
- App.jsx: 이벤트 수신 → Dialog 모달, router.navigate로 /login 이동
